Use long IDs; refactor mail, events and mappers
Change entity ID types to long/Long (User.id, Order.id, ownerId and related DTOs/repositories) and update repositories to use Long keys. Replace passing entities in async events with id-based payloads (OrderStatusChangedEvent) and adapt MailService/OrderStatusMailListener to accept orderId/userId, fetch entities via services, make MailService transactional, and improve recipient handling. Introduce and use request/response mappers in controllers and services (OrderRequestMapper, OrderResponseMapper, UserResponseMapper); refactor OrderService to return entities for controller mapping and adjust event publishing to send ids. Update tests and other callers to match the new ID types and API changes. Misc: minor null/format fixes (use String.valueOf for ids) and import/typing cleanups.

+++ src/main/java/de/hs_esslingen/besy/models/User.java
@@ -1,13 +1,21 @@
 package de.hs_esslingen.besy.models;
 
-import lombok.AllArgsConstructor;
-import lombok.RequiredArgsConstructor;
-import lombok.Builder;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
 import java.util.List;
+
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 
 @Getter
 @Setter
@@ -20,7 +28,7 @@ public class User {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private long id;
 
     @Column(name = "keycloak_uuid", nullable = true, unique = true)
     private String keycloakUUID;
@@ -40,7 +48,4 @@ public class User {
     @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
     private List<UserPreferences> preferences;
 
-
 }
-
-
